requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.23.0
numpy>=1.26.0
rank-bm25>=0.2.2
httpx>=0.27.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
duckduckgo-search>=5.2.0
numexpr>=2.8.8
pypdf>=4.2.0
beautifulsoup4>=4.12.0
trafilatura>=1.8.0
